refactor(cli): move onboarding inference validation helpers
