+++ test/gemini-probe-auth.test.ts
@@ -25,7 +25,7 @@ describe("Gemini dual-auth credential fix (issue #1960)", () => {
 
   describe("compiled probe uses ?key= for Gemini instead of Bearer header", () => {
     const onboardSrc = fs.readFileSync(
-      path.join(import.meta.dirname, "..", "dist", "lib", "onboard.js"),
+      path.join(import.meta.dirname, "..", "dist", "lib", "onboard-inference-validation.js"),
       "utf-8",
     );
